Change Guild constructor, change names in SimulationStates and BankAccount, update GuildTests

+++ 20190205/MiningSimulatorByKPMM/MiningSimulatorByKPMM/Locations/Guild/ICreateOreValue.cs
@@ -6,7 +6,7 @@ namespace MiningSimulatorByKPMM.Locations.Guild
 {
     public interface ICreateOreValue
     {
-        int GenerateSingleValue();
+        decimal GenerateSingleValue();
 
     }
 }
